pyproject.toml
[project]
name = "dwight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.5.0",
    "pydantic>=2.11.4",
]
